add kind filter to MCP search_code tool (code/example/documentation)
Moved ContentKind to src/types/chunk.rs so it's shared between the CLI
and MCP server. MCP search_code now accepts a `kinds` array matching
the CLI's -k/--type flag.
